+++ Xamarin.Forms.Controls.Issues/Xamarin.Forms.Controls.Issues.Shared/Issue8207.xaml
@@ -1,0 +1,33 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<local:TestShell xmlns="http://xamarin.com/schemas/2014/forms" xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml" x:Class="Xamarin.Forms.Controls.Issues.Issue8207" Title="Test Page" xmlns:local="using:Xamarin.Forms.Controls">
+    <ShellItem Title="Dashboard" FlyoutIcon="coffee.png" Route="Home">
+        <ShellContent Title="Control">
+            <ContentPage Title="Control">
+                <ScrollView>
+                    <StackLayout>
+                        <Label Text="Hello xamarin forms">
+                        </Label>
+                    </StackLayout>
+                </ScrollView>
+            </ContentPage>
+        </ShellContent>
+    </ShellItem>
+    <ShellItem Title="Connect" Route="connect">
+        <ShellSection>
+            <ShellContent Title="Connect">
+                <ContentPage Title="Connect" />
+            </ShellContent>
+            <ShellContent Title="Control">
+                <ContentPage Title="Control">
+                    <ScrollView>
+                        <StackLayout>
+                            <Label Text="Other Page">
+                            </Label>
+                        </StackLayout>
+                    </ScrollView>
+                </ContentPage>
+            </ShellContent>
+        </ShellSection>
+    </ShellItem>
+    <ShellContent Title="This is a shellContent" Icon="coffee.png" Route="dices" />
+</local:TestShell>
